style patch

+++ .config/hypr/scripts/gol-idle.sh
@@ -1,18 +1,24 @@
 #!/bin/bash
 
 PROGRAM="$HOME/.config/hypr/scripts/gol_wallpaper.sh"
-PGID_FILE="/tmp/hypridle_program.pgid"
+PGID_FILE="/tmp/gol_program.pid"
+FLAG_FILE="/tmp/gol_program.flag"
 
 case "$1" in
     idle)
+        touch "$FLAG_FILE"
         $PROGRAM &
         echo $(ps -o pgid= -p "$!") > "$PGID_FILE"
+        sleep 2
+        rm "$FLAG_FILE"
         ;;
     resume)
-        if [ -f "$PGID_FILE" ]; then
-            pgid=$(cat "$pgid_FILE")
-            pkill -g "$pgid"
-            rm "$pgid_FILE"
+        if [ ! -f "$FLAG_FILE" ]; then
+            if [ -f "$PGID_FILE" ]; then
+                pgid=$(cat "$PGID_FILE")
+                pkill -g "$pgid"
+                rm "$PGID_FILE"
+            fi
         fi
         ;;
     *)
